LESSONS LEARNED
Figured out why it kept yelling at me for why it didn't like croizats/xan's solution, it's because it pulled *-every-* single leve out of that sheet. And I need to filter out the crafters if I want to use the craftLeves sheet.... which in respospect works but ya know.

Hey! Atleast the test sheet is about 20fps improvement vs the old method (fucking bless). Time to see how much that stays when I wake up and add the rest of the data